xe: jit: gemm: improve Xe2 FHS strategy for transitionary m
Improves Xe2 strategy for trasitionary m in (32, 64]. The core issue
with the current strategies are that they do not have sufficient
paralleism to provide good performance accross a variety of n values.
The new strategies minimize the thread block size and add k-reduction by
a factor of 2 to provide increased performance.